requirements.txt
bottle==0.12.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Faker==19.13.0
python-dateutil==2.8.2 